s390x: vectorize crc32
This patch improves inflate performance by a factor of 15 on s390x.

Use vector extensions when compiling for s390x and binutils knows
about them. At runtime, check whether kernel supports vector
extensions (it has to be not just the CPU, but also the kernel) and
choose between the regular and the vectorized implementations.